# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowDeck productivity app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories for the authenticated user
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── GOALS ─────────────────────────────────────────────────────
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals for the authenticated user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /goals/{id}/progress:
    get:
      operationId: getGoalProgress
      tags: [goals]
      summary: Get progress detail for a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal progress detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalProgress"
        "404":
          description: Not found
  /goals/{id}/complete:
    post:
      operationId: completeGoal
      tags: [goals]
      summary: Mark a goal as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found

  # ── TASKS ─────────────────────────────────────────────────────
...
  /focus-sessions:
    get:
      operationId: listFocusSessions
      tags: [taskLogs]
      summary: List focus/Pomodoro sessions
      parameters:
        - name: taskId
          in: query
          required: false
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of focus sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FocusSession"
    post:
      operationId: createFocusSession
      tags: [taskLogs]
      summary: Log a focus/Pomodoro session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"

  # ── TAGS ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, userId, name, icon, color, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, icon, color]
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: string
        description:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: string
        description:
          type: string

    # ── GOAL ──────────────────────────────────────────────────
    Goal:
      type: object
      required: [id, userId, title, goalType, status, priority, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        goalType:
          type: string
          enum: [quantitative, habit, milestone]
        status:
          type: string
          enum: [active, paused, completed, archived]
        priority:
          type: string
          enum: [high, medium, low]
        targetValue:
          type: ["number", "null"]
        currentValue:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        targetEndDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    GoalInput:
      type: object
      required: [title, goalType, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        categoryId:
          type: integer
        goalType:
          type: string
          enum: [quantitative, habit, milestone]
        priority:
          type: string
          enum: [high, medium, low]
        targetValue:
          type: number
        startDate:
          type: string
        targetEndDate:
          type: string
        notes:
          type: string
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        categoryId:
          type: integer
        status:
          type: string
          enum: [active, paused, completed, archived]
        priority:
          type: string
          enum: [high, medium, low]
        targetValue:
          type: number
        currentValue:
          type: number
        startDate:
          type: string
        targetEndDate:
          type: string
        notes:
          type: string
    GoalProgress:
      type: object
      required: [goalId, percent, currentValue, targetValue]
      properties:
        goalId:
          type: integer
        percent:
          type: number
        currentValue:
          type: number
        targetValue:
          type: number
        completedMilestones:
          type: integer
        totalMilestones:
          type: integer

    # ── TASK ──────────────────────────────────────────────────
...
    FocusSession:
      type: object
      required: [id, userId, durationMinutes, sessionDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        taskId:
          type: ["integer", "null"]
        durationMinutes:
          type: integer
        sessionType:
          type: string
          enum: [pomodoro, short_break, long_break, custom]
        sessionDate:
          type: string
        createdAt:
          type: string
    FocusSessionInput:
      type: object
      required: [durationMinutes, sessionDate]
      properties:
        taskId:
          type: integer
        durationMinutes:
          type: integer
        sessionType:
          type: string
          enum: [pomodoro, short_break, long_break, custom]
        sessionDate:
          type: string

    # ── TAG ───────────────────────────────────────────────────
...