# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDeck productivity app API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalInputGoalType } from "./goalInputGoalType";
import type { GoalInputPriority } from "./goalInputPriority";

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  categoryId?: number;
  goalType: GoalInputGoalType;
  priority: GoalInputPriority;
  targetValue?: number;
  startDate?: string;
  targetEndDate?: string;
  notes?: string;
}