# lib/api-zod/src/generated/types/goalProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowDeck productivity app API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalProgress {
  goalId: number;
  percent: number;
  currentValue: number;
  targetValue: number;
  completedMilestones?: number;
  totalMilestones?: number;
}